+++ CHANGELOG.rst
@@ -23,6 +23,10 @@ Changelog
 * Added initial support for parsing PKCS12 files with
   :func:`~cryptography.hazmat.primitives.serialization.pkcs12.load_key_and_certificates`.
 * Added support for :class:`~cryptography.x509.IssuingDistributionPoint`.
+* Added `rfc4514_string()` method to :class:`~cryptography.x509.Name`,
+  :class:`~cryptography.x509.RelativeDistinguishedName` and
+  :class:`~cryptography.x509.NameAttribute` to format the name or component as
+  a RFC 4514 Distinguished Name string.
 
 .. _v2-4-2: